test(es/typescript): Move namespace fold coverage to fixtures
`AGENTS.md` asks for fixture tests over inline ones and for new coverage to
extend the existing suites, and the sibling work in #12101 and #12179 is
already under `tests/fixture/issue-*`. Move the three default-config cases
there; the output is byte-identical to the inline snapshots they replace.

The `tsEnumIsMutable` case stays inline: the fixture harness runs a fixed
pass and has no way to pass a non-default config, which is why
`ts_enum_is_mutable_true` is inline too.
