+++ crates/swc_ecma_transforms_typescript/tests/fixture/issue-12102-namespace-member/output.js
