+++ crates/swc_ecma_transforms_typescript/tests/fixture/issue-12102-stays-opaque/output.js
