+++ crates/swc_ecma_transforms_typescript/tests/strip.rs
@@ -2919,132 +2919,3 @@ test!(
     ts_jsx_bad_pragma,
     r#"/** @jsx bad-pragma */"#
 );
-
-test!(
-    Syntax::Typescript(TsSyntax::default()),
-    tr,
-    ts_enum_namespace_member,
-    r#"
-    namespace N { export const foo = "s"; }
-    enum E { A = N.foo, B = "b" }
-
-    namespace M { export namespace Inner { export const bar = "t"; } }
-    enum F { A = M.Inner.bar, B = "b" }
-
-    namespace C1 { export namespace C2 { export namespace C3 { export const deep = "d"; } } }
-    enum G { A = C1.C2.C3.deep, B = "b" }
-
-    namespace Dot.Sub { export const x = "dotted"; }
-    enum H { A = Dot.Sub.x, B = "b" }
-
-    enum Merged { foo = 1 }
-    namespace Merged { export const bar = "merged"; }
-    enum I { A = Merged.foo, B = Merged.bar, C = "c" }
-
-    namespace WithEnum { export enum Inner { X = 1 } }
-    enum J { A = WithEnum.Inner.X, B = "b" }
-
-    namespace SelfQ { export const a = "x"; export enum E { A = SelfQ.a, B = "b" } }
-    namespace OuterQ { export const a = "y"; export namespace Inner { export enum E { A = OuterQ.a, B = "b" } } }
-
-    namespace ConDE { export declare enum DE { X = 3 } }
-    enum K { A = ConDE.DE.X, B = "b" }
-
-    namespace Opt { export namespace I { export const x = "s"; } }
-    enum L { A = Opt?.I.x, B = "b" }
-    enum M2 { A = Opt.I?.x, B = "b" }
-    enum N2 { A = Opt?.I?.x, B = "b" }
-
-    namespace Opt3 { export namespace H { export namespace K { export const y = "t"; } } }
-    enum O2 { A = Opt3?.H?.K.y, B = "b" }
-    "#
-);
-
-test!(
-    Syntax::Typescript(TsSyntax::default()),
-    tr,
-    ts_enum_ambient_namespace_member,
-    r#"
-    declare namespace D { enum A { X = 1 } }
-    enum E { P = D.A.X, Q }
-
-    declare namespace B1 { const enum K { Z = 3 } }
-    enum F { P = B1.K.Z, Q }
-
-    declare namespace AC { export const c = 7; }
-    enum G { P = AC.c, Q }
-
-    declare namespace AD { namespace Inner { enum E { X = 1 } } }
-    enum H { P = AD.Inner.E.X, Q }
-
-    declare namespace AE { export namespace Inner { export const c = 5; } }
-    enum I { P = AE.Inner.c, Q }
-
-    declare namespace AF.Dotted { const k = 9; }
-    enum J { P = AF.Dotted.k, Q }
-
-    namespace Con { export declare namespace Amb { const v = 1; } }
-    enum K { P = Con.Amb.v, Q }
-
-    declare namespace Amb2 { namespace Inner { const w = 2; } }
-    enum L { P = Amb2.Inner.w, Q }
-    "#
-);
-
-test!(
-    Syntax::Typescript(TsSyntax::default()),
-    tr,
-    ts_enum_namespace_member_stays_opaque,
-    r#"
-    enum Fwd { A = Later.baz, B = "b" }
-    namespace Later { export const baz = "u"; }
-
-    namespace NotExported { const hidden = "h"; }
-    enum Hid { A = NotExported.hidden, B = "b" }
-
-    namespace Mut { export let m = "v"; }
-    enum MutE { A = Mut.m, B = "b" }
-
-    const objLit = { k: "w" } as const;
-    enum Obj { A = objLit.k, B = "b" }
-
-    declare namespace D2 { const c: string; }
-    enum H { A = D2.c, B = "b" }
-
-    namespace T { export const typed: string = "annotated"; }
-    enum Typed { A = T.typed, B = "b" }
-
-    namespace Outer { namespace Hidden { export const v = "h"; } }
-    enum Nested { A = Outer.Hidden.v, B = "b" }
-
-    namespace WithHidden { export const live = 0; declare const hidden = 1; }
-    enum FalseAmbient { A = WithHidden.hidden, B = "b" }
-
-    enum LateConst { A = LaterNs.b, B = "b" }
-    namespace LaterNs { export const b = "post"; }
-
-    enum LateEnum { A = LaterEnumNs.Inner.X, B = "b" }
-
-    namespace Elem { export namespace Seg { export const x = 1; } export const y = 2; }
-    enum ComputedSeg { A = Elem["Seg"].x, B = "b" }
-    enum ComputedProp { A = Elem["y"], B = "b" }
-    enum TplSeg { A = Elem[`Seg`].x, B = "b" }
-
-    namespace Paren { export const p = "p"; }
-    enum ParenObj { A = (Paren).p, B = "b" }
-
-    declare global { const g = 1; }
-    const global = { g: 99 };
-    enum GlobalObj { A = global.g, B = "b" }
-
-    namespace PrivAmb { export const live = 0; declare enum HiddenE { X = 1 } }
-    enum HiddenEnum { A = PrivAmb.HiddenE.X, B = "b" }
-
-    namespace PrivNs { export const live = 0; declare namespace HiddenN { const x = "s"; } }
-    enum HiddenNs { A = PrivNs.HiddenN.x, B = "b" }
-
-    namespace PrivC { export const live = 0; declare const enum HiddenC { Z = 3 } }
-    enum HiddenConst { A = PrivC.HiddenC.Z, B = "b" }
-    namespace LaterEnumNs { export enum Inner { X = 1 } }
-    "#
-);
